Dynamic routes: PROBLEM
Cant get gridsome to bundle the brand pages from a dynamic route.
I know it would be easier to write the four brand pages and let it be, but I want it to be a dynamic route that pull the info from my data file.
Gridsome stop for now, exploring VuePress and Nuxt...